re-align customauth
This PR seeks to re-align the current implementation of custom-auth-swift to that of the web implementation of CustomAuth.

This will reduce the overhead of differences a developer would need to remember between the two and provide an overall better experience.

This will also provide better long term maintainability and make feature parity simpler to implement between the two.

Note this is a breaking change and needs to be released as a major version.